requirements.txt
ansible-vault
logging
pandas
scikit-learn
pyyaml
nltk
numpy
coverage
psycopg2-binary